🐛 二回目のログインでユーザーの情報が上書きされてしまっていたバグの修正
